refactor: rename and extract podcast tests
