+++ frontend/tests/podcast/homepage/podcast.category.spec.ts
@@ -1,0 +1,80 @@
+import test, { expect, Page } from "@playwright/test"
+import { allPodcastCategories } from "../../mocks/podcast.category"
+import { HOMEPAGE } from "../../constants/homepageConstants"
+
+test.describe("Podcast Homepage /podcasts", () => {
+  test.describe("Podcast Categories Section", () => {
+    function getRefreshPodcastCategoryButton(page: Page) {
+      return page.locator(".podcast-category-container").getByRole("button", {
+        name: "refresh podcast categories",
+        exact: true,
+      })
+    }
+
+    test("should display podcast categories in slider", async ({ page }) => {
+      await page.route("*/**/api/podcast/category", async (route) => {
+        const json = allPodcastCategories
+        await route.fulfill({ json })
+      })
+      await page.goto(HOMEPAGE + "/podcasts")
+      await expect(page).toHaveTitle(/xtal - podcasts/)
+      await expect(page.locator(".podcast-category-container")).toBeVisible()
+      await expect(
+        page.locator(".podcast-category-container .podcast-category-title")
+      ).toBeVisible()
+      for (const category of allPodcastCategories.data) {
+        await expect(
+          page
+            .locator(".podcast-category-slider")
+            .getByText(category.name, { exact: true }),
+          `Podcast category "${category.name}" should be visible`
+        ).toBeVisible()
+      }
+    })
+
+    test("should display error message and refresh category button when podcast categories cannot be fetched", async ({
+      page,
+    }) => {
+      await page.route("*/**/api/podcast/category", async (route) => {
+        const json = []
+        await route.fulfill({ json })
+      })
+      await page.goto(HOMEPAGE + "/podcasts")
+      await expect(page).toHaveTitle(/xtal - podcasts/)
+      await expect(
+        page
+          .locator(".podcast-category-container")
+          .getByText(
+            "Could not get podcast categories. Please try again later",
+            { exact: true }
+          )
+      ).toBeVisible()
+      await expect(getRefreshPodcastCategoryButton(page)).toBeVisible()
+      await expect(page.getByTestId("loading-spinner")).not.toBeVisible()
+    })
+
+    test("should refresh podcast categories on button click", async ({
+      page,
+    }) => {
+      let shouldFetchData = false
+      await page.route("*/**/api/podcast/category", async (route) => {
+        const json = shouldFetchData ? allPodcastCategories : []
+        await route.fulfill({ json })
+      })
+      await page.goto(HOMEPAGE + "/podcasts")
+      await expect(page).toHaveTitle(/xtal - podcasts/)
+      expect(getRefreshPodcastCategoryButton(page)).toBeVisible()
+
+      shouldFetchData = true
+      await getRefreshPodcastCategoryButton(page).click()
+      for (const category of allPodcastCategories.data) {
+        await expect(
+          page
+            .locator(".podcast-category-slider")
+            .getByText(category.name, { exact: true }),
+          `Podcast category "${category.name}" should be visible`
+        ).toBeVisible()
+      }
+    })
+  })
+})
