+++ frontend/tests/podcast/homepage/podcast.trending.spec.ts
@@ -1,149 +1,13 @@
 import dayjs from "dayjs"
 import test, { expect, Page } from "@playwright/test"
 import {
-  clickRandomRadioStationButton,
-  getNavbarPodcastLink,
-  getNavbarRadioLink,
-  getRadioCardMapPopup,
-  getRadioStationMapPopupCloseButton,
-  getToastMessages,
-  HOMEPAGE,
-} from "../constants/homepageConstants"
-import {
-  getFavouriteStationsButton,
-  getFavouriteStationsDrawer,
-  getRadioCardFavouriteIcon,
-} from "../constants/favouriteStationConstants"
-import { unitedStatesStation } from "../mocks/station"
-import {
   defaultTenTrendingPodcasts,
   threeTrendingPodcasts,
   zeroTrendingPodcasts,
-} from "../mocks/podcast.trending"
-import { allPodcastCategories } from "../mocks/podcast.category"
+} from "../../mocks/podcast.trending"
+import { getToastMessages, HOMEPAGE } from "../../constants/homepageConstants"
 
 test.describe("Podcast Homepage /podcasts", () => {
-  test("should display title", async ({ page }) => {
-    await page.goto(HOMEPAGE + "/podcasts")
-    await expect(page).toHaveTitle(/xtal - podcasts/)
-  })
-
-  test("should navigate back to homepage (/) header navbar radio link is clicked", async ({
-    page,
-  }) => {
-    await page.goto(HOMEPAGE + "/podcasts")
-    await expect(page).toHaveTitle(/xtal - podcasts/)
-    expect(page.url()).toMatch(/\/podcasts$/)
-    await getNavbarRadioLink(page).click()
-    await expect(page).not.toHaveTitle(/xtal - podcasts/)
-    expect(page.url()).not.toMatch(/\/podcasts$/)
-  })
-
-  test("should load favourite station and navigate back to homepage when load station button is clicked in favourite stations drawer", async ({
-    page,
-  }) => {
-    await page.route("*/**/json/stations/search?*", async (route) => {
-      const json = [unitedStatesStation]
-      await route.fulfill({ json })
-    })
-    await page.goto(HOMEPAGE)
-    await clickRandomRadioStationButton(page)
-    await expect(getRadioCardFavouriteIcon(page)).toBeVisible()
-    await getRadioCardFavouriteIcon(page).click()
-    await getRadioStationMapPopupCloseButton(page).click()
-
-    await getNavbarPodcastLink(page).click()
-    await getFavouriteStationsButton(page).click()
-    await getFavouriteStationsDrawer(page)
-      .locator(".favourite-station")
-      .getByRole("button", {
-        name: "load station",
-      })
-      .click()
-    await expect(
-      getRadioCardMapPopup(page).getByRole("heading", {
-        name: unitedStatesStation.name,
-        exact: true,
-      })
-    ).toBeVisible()
-    expect(page.url()).not.toMatch(/\/podcasts$/)
-  })
-
-  test.describe("Podcast Categories Section", () => {
-    function getRefreshPodcastCategoryButton(page: Page) {
-      return page.locator(".podcast-category-container").getByRole("button", {
-        name: "refresh podcast categories",
-        exact: true,
-      })
-    }
-
-    test("should display podcast categories in slider", async ({ page }) => {
-      await page.route("*/**/api/podcast/category", async (route) => {
-        const json = allPodcastCategories
-        await route.fulfill({ json })
-      })
-      await page.goto(HOMEPAGE + "/podcasts")
-      await expect(page).toHaveTitle(/xtal - podcasts/)
-      await expect(page.locator(".podcast-category-container")).toBeVisible()
-      await expect(
-        page.locator(".podcast-category-container .podcast-category-title")
-      ).toBeVisible()
-      for (const category of allPodcastCategories.data) {
-        await expect(
-          page
-            .locator(".podcast-category-slider")
-            .getByText(category.name, { exact: true }),
-          `Podcast category "${category.name}" should be visible`
-        ).toBeVisible()
-      }
-    })
-
-    test("should display error message and refresh category button when podcast categories cannot be fetched", async ({
-      page,
-    }) => {
-      await page.route("*/**/api/podcast/category", async (route) => {
-        const json = []
-        await route.fulfill({ json })
-      })
-      await page.goto(HOMEPAGE + "/podcasts")
-      await expect(page).toHaveTitle(/xtal - podcasts/)
-      await expect(
-        page
-          .locator(".podcast-category-container")
-          .getByText(
-            "Could not get podcast categories. Please try again later",
-            { exact: true }
-          )
-      ).toBeVisible()
-      await expect(getRefreshPodcastCategoryButton(page)).toBeVisible()
-      await expect(page.getByTestId("loading-spinner")).not.toBeVisible()
-    })
-
-    test("should refresh podcast categories on button click", async ({
-      page,
-    }) => {
-      let shouldFetchData = false
-      await page.route("*/**/api/podcast/category", async (route) => {
-        const json = shouldFetchData ? allPodcastCategories : []
-        await route.fulfill({ json })
-      })
-      await page.goto(HOMEPAGE + "/podcasts")
-      await expect(page).toHaveTitle(/xtal - podcasts/)
-      expect(getRefreshPodcastCategoryButton(page)).toBeVisible()
-
-      shouldFetchData = true
-      await getRefreshPodcastCategoryButton(page).click()
-      for (const category of allPodcastCategories.data) {
-        await expect(
-          page
-            .locator(".podcast-category-slider")
-            .getByText(category.name, { exact: true }),
-          `Podcast category "${category.name}" should be visible`
-        ).toBeVisible()
-      }
-    })
-  })
-
   test.describe("Trending Podcasts Section", () => {
     function getPodcastCards(page: Page) {
       return page.locator(".podcast-trending-container .podcast-trending-card")
